Lots of fixes for Mac version
git-svn-id: https://wiic.svn.sourceforge.net/svnroot/wiic@12 fe444ef7-8870-402c-b3f2-cc32b5d63832

+++ src/example/example.cpp
@@ -428,7 +428,7 @@ int main(int argc, char** argv)
     // Connect to the wiimote
     std::vector<CWiimote>& wiimotes = wii.Connect();
 
-    printf("Connected to %u wiimotes\n", wiimotes.size());
+    printf("Connected to %u wiimotes\n", (unsigned int)wiimotes.size());
 
     // Setup the wiimotes
     for(index = 0, i = wiimotes.begin(); i != wiimotes.end(); ++i, ++index)
